Feat: couponTable 분리
